% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEpivizBlocks.R
\name{convertEpivizBlocks}
\alias{convertEpivizBlocks}
\title{Convert Epiviz BlocksTrack to Gviz AnnotationTrack.}
\usage{
convertEpivizBlocks(app, chart_obj, chr)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{chart_obj}{(EpivizChart) an object of class \code{\link[epivizr]{EpivizChart}}.}

\item{chr}{(character) the name of the chromosome to plot over, ex: "chr11".}
}
\value{
A list containing an object of class \code{\link[Gviz]{AnnotationTrack}}
}
\description{
Convert Epiviz BlocksTrack to Gviz AnnotationTrack.
}
\examples{
\dontrun{
# see package vignette for example usage
convertEpivizBlocks(app, chart_obj, chr)
}

}

