% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEpivizGenes.R
\name{convertEpivizGenes}
\alias{convertEpivizGenes}
\title{Convert Epiviz GenesTrack to Gviz GeneRegionTrack.}
\usage{
convertEpivizGenes(app, chart_obj, chr = NULL)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{chart_obj}{(EpivizChart) an object of class \code{\link[epivizr]{EpivizChart}}.}

\item{chr}{(character) the name of the chromosome to plot over, ex: "chr11".}
}
\value{
A list containing an object of class \code{\link[Gviz]{IdeogramTrack}}, \code{\link[Gviz]{GenomeAxisTrack}}, \code{\link[Gviz]{GeneRegionTrack}}
}
\description{
Convert Epiviz GenesTrack to Gviz GeneRegionTrack.
}
\examples{
# see package vignette for example usage
convertEpivizGenes(app, chart_obj, chr=NULL)

}

