% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEpivizLine.R
\name{convertEpivizLine}
\alias{convertEpivizLine}
\title{Convert Epiviz LineTrack to Gviz DataTrack.}
\usage{
convertEpivizLine(app, chart_obj, chr)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{chart_obj}{(EpivizChart) an object of class \code{\link[epivizr]{EpivizChart}}.}

\item{chr}{(character) the name of the chromosome to plot over, ex: "chr11".}
}
\value{
A list containing an object of class \code{\link[Gviz]{DataTrack}}
}
\description{
Convert Epiviz LineTrack to Gviz DataTrack.
}
\examples{
# see package vignette for example usage
\dontrun{
convertEpivizLine(app, chart_obj, chr)
}

}

