% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEpivizStackedLine.R
\name{convertEpivizStackedLine}
\alias{convertEpivizStackedLine}
\title{Convert Epiviz StackedLineTrack to Gviz DataTrack.}
\usage{
convertEpivizStackedLine(app, chart_obj, chr)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{chart_obj}{(EpivizChart) an object of class \code{\link[epivizr]{EpivizChart}}.}

\item{chr}{(character) the name of the chromosome to plot over, ex: "chr11".}
}
\value{
A list containing an object of class \code{\link[Gviz]{DataTrack}}
}
\description{
Convert Epiviz StackedLineTrack to Gviz DataTrack.
}
\examples{
# see package vignette for example usage
\dontrun{
convertEpivizStackedLine(app, chart_obj, chr)
}

}

