% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGvizAnnotation.R
\name{convertGvizAnnotation}
\alias{convertGvizAnnotation}
\title{Convert Gviz AnnotationTrack to Epiviz BlocksTrack.}
\usage{
convertGvizAnnotation(app, track)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{track}{(AnnotationTrack) an object of class \code{\link[Gviz]{AnnotationTrack}}.}
}
\value{
An object of class \code{\link[epivizr]{EpivizChart}}
}
\description{
Convert Gviz AnnotationTrack to Epiviz BlocksTrack.
}
\examples{
# see package vignette for example usage
convertGvizAnnotation(app, track)

}

