% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGvizData.R
\name{convertGvizData}
\alias{convertGvizData}
\title{Convert Gviz DataTrack to Epiviz LineTrack.}
\usage{
convertGvizData(app, track)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{track}{(DataTrack) an object of class \code{\link[Gviz]{DataTrack}}.}
}
\value{
An object of class \code{\link[epivizr]{EpivizChart}}
}
\description{
Convert Gviz DataTrack to Epiviz LineTrack.
}
\examples{
# see package vignette for example usage
\dontrun{
convertGvizData(app, track)
}

}

