% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGvizGeneRegion.R
\name{convertGvizGeneRegion}
\alias{convertGvizGeneRegion}
\title{Convert Gviz GeneRegionTrack to Epiviz GenesTrack.}
\usage{
convertGvizGeneRegion(track, annotation = NULL, chr = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{track}{(GeneRegionTrack) an object of class \code{\link[Gviz]{GeneRegionTrack}}.}

\item{annotation}{(OrganismDb) an object of class \code{\link[OrganismDbi]{OrganismDb}}.}

\item{chr}{(character) chromosome to browse to on app startup.}

\item{start}{(integer) start location to browse to on app startup.}

\item{end}{(integer) end location to browse to on app startup.}
}
\value{
An object of class \code{\link[epivizr]{EpivizApp}}
}
\description{
Convert Gviz GeneRegionTrack to Epiviz GenesTrack.
}
\examples{
# see package vignette for example usage
\dontrun{
convertGvizGeneRegion(track, annotation=NULL, chr=NULL, start=NULL, end=NULL)
}

}

