% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizToGviz.R
\name{epivizToGviz}
\alias{epivizToGviz}
\title{Convert an existing Epiviz workspace into a Gviz plot.}
\usage{
epivizToGviz(app, plot_tracks = TRUE)
}
\arguments{
\item{app}{(EpivizApp) an object of class \code{\link[epivizr]{EpivizApp}}.}

\item{plot_tracks}{(logical) plot all of the Gviz tracks together.}
}
\value{
A list of GdObjects that represent Gviz tracks, ex: \code{\link[Gviz]{GeneRegionTrack}}, \code{\link[Gviz]{AnnotationTrack}}, \code{\link[Gviz]{DataTrack}}
}
\description{
Convert an existing Epiviz workspace into a Gviz plot.
}
\examples{
# see package vignette for example usage
epivizToGviz(app, plot_tracks=TRUE)

}

