% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvizToEpiviz.R
\name{gvizToEpiviz}
\alias{gvizToEpiviz}
\title{Convert an existing Gviz plot into an Epiviz workspace.}
\usage{
gvizToEpiviz(gviz_plot, annotation = NULL, chr = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{gviz_plot}{(list) a list representing a plot from \code{\link[Gviz]{plotTracks}}}

\item{annotation}{(OrganismDb) an object of class \code{\link[OrganismDbi]{OrganismDb}}.}

\item{chr}{(character) chromosome to browse to on app startup.}

\item{start}{(integer) start location to browse to on app startup.}

\item{end}{(integer) end location to browse to on app startup.}
}
\value{
An object of class \code{\link[epivizr]{EpivizApp}}
}
\description{
Convert an existing Gviz plot into an Epiviz workspace.
}
\examples{
# see package vignette for example usage
gvizToEpiviz(gviz_plot, annotation=NULL, chr=NULL, start=NULL, end=NULL)

}

